requirements.txt
aws-cdk-lib==2.130.0
aws-cdk.integ-tests-alpha==2.130.0a0
cdk-integ-runner-cwd-fix==0.1.1
constructs>=10.0.0,<11.0.0
boto3==1.34.54

black==24.2.0
isort==5.13.2
